requirements.txt
bs4==0.0.1
numpy==1.19.5
pandas==1.3.0
IPython==7.25.0
matplotlib==3.3.4
pyyaml==5.4.1
